Fire a chosen:update event after changing the result box.
Rationale: sometimes css is just not enough to layout things, especially
with dynamically chaning content. The update event is fired

______AFTER______

the result box has been updated and the pull-down/up menu has been
hidden. At that time, there is some potential to fetch reliable values
for the dimension of the chosen container.